Feature/simulations

* Add the "simulate" command

    The "simulate" command ("config simulate") allows one to run a
    calculation without running the fit procedure. The results are given in
    the form of text files (the usual ".dat" files) and standard plot files
    (the "Plots" folder) minus the files related to the experiment files.

    Example:

    chemex simulate -e Experiments/*.toml -p Parameters/*.toml -o Output

* Fix bug and add example for "simulate"

    Add the script "simulate.sh" to show how to run a simulation in
    "examples/CEST_15N".

    Fix a bug issuing a warning when simulating CEST profiles.

* Change the behaviour of default param values

    When a parameter is constrained the default value given by the user in
    the parameter.toml is ignored; the constraint remains and is used to
    define the parameter value.

    Re issue #19.

* Fix bug related to plotting

+++ chemex/containers/experiment.py
@@ -60,9 +60,9 @@ class Experiments:
         for experiment in self._experiments.values():
             experiment.write(params, path_)
 
-    def plot(self, params, path):
+    def plot(self, params, path, simulation=False):
         for experiment in self._experiments.values():
-            experiment.plot(params, path)
+            experiment.plot(params, path, simulation)
 
     def select(self, selection=None):
         for experiment in self._experiments.values():
@@ -142,7 +142,7 @@ class RelaxationExperiment:
             self.filename, self.exp_type, profiles=profiles, verbose=False
         )
 
-    def plot(self, params, path):
+    def plot(self, params, path, simulation=False):
         basename = path / self.filename.name
         name_pdf = basename.with_suffix(".pdf")
         name_exp = basename.with_suffix(".exp")
@@ -150,11 +150,13 @@ class RelaxationExperiment:
         print(f"  - {name_pdf} [.fit, .exp]")
         with cl.ExitStack() as stack:
             file_pdf = stack.enter_context(pdf.PdfPages(str(name_pdf)))
-            file_exp = stack.enter_context(name_exp.open("w"))
             file_fit = stack.enter_context(name_fit.open("w"))
+            if not simulation:
+                file_exp = stack.enter_context(name_exp.open("w"))
+            else:
+                file_exp = None
             for profile in sorted(self._profiles):
-                profile.plot(params, file_pdf)
-                profile.write_plot(params, file_exp, file_fit)
+                profile.plot(params, file_pdf, file_exp, file_fit, simulation)
 
     def write(self, params, path):
         filename = (path / self.filename.name).with_suffix(".dat")
